requirements.txt
fastapi
uvicorn
chromadb
pydantic
pytest
requests
python-dotenv
sqlalchemy
openai>=1.25.0
langgraph>=0.0.40
langchain>=0.2.0
tiktoken
defusedxml
psycopg2-binary
pymupdf>=1.23.0
aiohttp>=3.9.0
networkx>=3.0
python-jose[cryptography]